Fix of General error: 3780 Referencing column

+++ database/migrations/2020_06_20_010403_create_add_publish_post_collumn_table.php
@@ -14,7 +14,7 @@ class CreateAddPublishPostCollumnTable extends Migration
     public function up()
     {
         Schema::table('Posts', function (Blueprint $table) {
-            $table->integer('published')->default(0);
+            $table->bigInteger('published')->default(0);
         });
     }
